Feat: 관리자, 작성자 삭제권한 추가

+++ legolego/src/main/java/com/kosta/legolego/diypackage/controller/DiyController.java
@@ -81,9 +81,20 @@ public class DiyController {
 
   }
   @DeleteMapping("/user/packages/{package_num}")
-  //추후 삭제 권한 검사 추가
-  public ResponseEntity<?> deleteDiy(@PathVariable("package_num") Long package_num) {
+  public ResponseEntity<?> deleteDiy(@PathVariable("package_num") Long package_num, @AuthenticationPrincipal CustomUserDetails userDetails) {
     try{
+      if (userDetails == null || !userDetails.getRole().equals("ROLE_USER")) {
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+      }
+      Long userNum = userDetails.getId();
+      //요청된 package_num에 해당하는 DIY 패키지의 작성자 확인
+      Long packageOwner = diyService.getPackageOwner(package_num);
+
+      //작성자와 인증된 사용자 ID 비교
+      if (!userNum.equals(packageOwner)) {
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 작성자가 아닌 경우 403 Forbidden 반환
+      }
+
       diyService.deleteDiy(package_num);
       return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }catch (RuntimeException e){
